Router argument issue
I don't generate any routes with herbert.  I noticed that my nginx error log was getting flooded with errors about foreach not having a valid argument. So I added a check to make sure that there is in fact an array of routes to add before trying to iterate it! No more log spam.